requirements.txt
Django==1.9.7
djangorestframework==3.3.3
django-extensions==1.5.9
django-debug-toolbar==1.4